Add more Windows event debug logging to the API. (st) Determine the correct widget to use in Windows. (st) Fix API tty table extension. (st)
git-svn-id: svn://mielke.cc/main/brltty@1683 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/scr.h
@@ -34,7 +34,7 @@ typedef enum {
 typedef struct {
   short rows, cols;	/* screen dimensions */
   short posx, posy;	/* cursor position */
-  short no;		      /* screen number */
+  int no;		      /* screen number */
 } ScreenDescription;
 
 typedef struct {
